Update documentation and function names
